[libnozzle] add api_nozzle_open test
